Code snippet for perfectly centering text in divs (Horizontally and Vertically)
Here is the complete code snippet for this tutorial (with browser fallbacks). If you're wondering about the transform:translate property-value pair, the translateY value affects the Y-axis movement (up and down), whereas the translateX value affects the X-axis movement (left and right). The translateZ value affects our zoom-in and zoom-out perspective (inward and outwards).