Uploaded code and tutorial on creating flipping cards
This code snippet and accompanying tutorial shows how to flip an HTML element like a box or div and display content on the reverse side, a great space saver and booster of UI / UX scores for websites. Using pure CSS, the code contains no JavaScript and therefore will likely retain a high search score for website rankings and indexing.